Service interface 생성
