+++ src/main/java/j2kb/ponicon/scrap/mypage/MyPageController.java
@@ -3,10 +3,8 @@ package j2kb.ponicon.scrap.mypage;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiParam;
-import j2kb.ponicon.scrap.category.dto.GetCategoryListRes;
 import j2kb.ponicon.scrap.mypage.dto.GetUserRes;
 import j2kb.ponicon.scrap.response.BaseResponse;
-import j2kb.ponicon.scrap.user.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -18,7 +16,7 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("/user")
 @RequiredArgsConstructor
 public class MyPageController {
-    private final MyPageService myPageService;
+    private final MyPageServiceImpl myPageService;
     /**
      * 카테고리 조회 API
      * UserId를 RequestParam으로 받아서 myPageService.userInfo 후 마이페이지를 조회하는 역할을 합니다.
